careerguidancemodel/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-generativeai==0.8.0
requests==2.32.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
